+++ testing_suite/unit_tests/e_announcements_tests/campaign_controller_tests/test_is_date_friday_before_easter.py
@@ -17,9 +17,9 @@ class IsDateFridayBeforeEasterTestCase(CampaignControllerBaseTestCase):
 
     def test_is_date_friday_before_easter(self):
         friday_before_easter = datetime(2017, 4, 14)
-        response = self.controller.is_date_friday_before_easter(friday_before_easter)
+        response = self.base.is_date_friday_before_easter(friday_before_easter)
         self.assertTrue(response)
 
         friday_after_easter = datetime(2017, 4, 21)
-        response = self.controller.is_date_friday_before_easter(friday_after_easter)
+        response = self.base.is_date_friday_before_easter(friday_after_easter)
         self.assertFalse(response)
