+++ tinker/e_announcements/e_announcements_controller.py
@@ -1,6 +1,7 @@
 # Global
 from datetime import datetime
 from bu_cascade.asset_tools import find
+import math
 
 # Packages
 from bu_cascade.asset_tools import update
@@ -256,3 +257,52 @@ class EAnnouncementsController(TinkerController):
                 if second_ea_date == str(date_id):
                     self.get_title_and_message(form, ea_display)
         return ea_display
+
+    def is_bethel_holiday(self, date):
+        # New years
+        if date.month == 1 and date.day == 1:
+            return True
+        # New Years(observed) -- If new years day is on the weekend, we get the monday off (2nd or 3rd)
+        elif date.month == 1 and date.weekday() == 0 and (date.day == 2 or date.day == 3):
+            return True
+        # MLK Day - 3rd monday in jan
+        elif date.month == 1 and date.weekday() == 0 and math.ceil(date.day/7.0) == 3:
+            return True
+        # Easter (is the date the friday before easter)
+        elif self.is_date_friday_before_easter(date):
+            return True
+        # memorial day - last monday in may (may, date is after 24th and its a monday)
+        elif date.month == 5 and date.day > 24 and date.weekday() == 0:
+            return True
+        # july 4
+        elif date.month == 7 and date.day == 4:
+            return True
+        # Labor Day - first monday in sept
+        elif date.month == 9 and date.weekday() == 0 and math.ceil(date.day/7.0) == 1:
+            return True
+        # Black Friday -- the friday after the fourth thursday in nov
+        elif date.month == 11 and math.ceil((date.day-1)/7.0) == 4 and date.weekday() == 4:
+            return True
+        # Christmas Eve(observed) - christmas eve is on the weekend, we get the friday off (22nd or 23rd).
+        elif date.month == 12 and date.weekday() == 4 and (date.day == 22 or date.day == 23):
+            return True
+        # christmas days
+        elif date.month == 12 and date.day >= 24:
+            return True
+
+        return False
+
+    def is_date_friday_before_easter(self, date):
+        year = date.year
+
+        # code from http://code.activestate.com/recipes/576517-calculate-easter-western-given-a-year/
+        a = year % 19
+        b = year // 100
+        c = year % 100
+        d = (19 * a + b - b // 4 - ((b - (b + 8) // 25 + 1) // 3) + 15) % 30
+        e = (32 + 2 * (b % 4) + 2 * (c // 4) - d - (c % 4)) % 7
+        f = d + e - 7 * ((a + 11 * d + 22 * e) // 451) + 114
+        month = f // 31
+        day = f % 31 + 1
+
+        return date.month == month and date.day == (day - 2)
